feat: track winner with primes
I used prime factorisation to assign each winning state of the game board a unique numerical value. Each round of the game, it is possible to mark off which spots of the array are marked-off then filter that array to extract only the elements that are numbers. From there, it is possible to multiply the whole array together. If the resulting number marks off a unique state of the game board, so it can be used to determine if the game has been won yet or not (by comparing it to the array that stores all the numerical values for winning states of the board, and returning if a match is found with the actual state of the board as stored in the xChecker or oChecker arrays).