feat: automate turns
I automated the turns. Turns are controlled by a counter. When the count is divisible by 2, it is playerOne's turn; when the count is not divisible by 2, it is playerTwo's turn. The counter is only incremented if the player successfully marks off a spot on the game board (so if they try to mark off a spot that is already taken, the counter isn't incremented so the player can try again)