feat: win, spot taken animations
I have added animations that display when a player has won a game of tic-tac-toe, and animations that display when a player tries to take a spot on the game board that has already been taken.

The animations are as follows.

When a player wins, the background flashes and the contents of the screen pulse in unison. The winning spots on the game board pulse more than all the other elements, and they flash different colours, too.

As per taken animation, I've made it so that when a user tries to click on a spot on the game board that is taken, the text in that spot will shake and turn red. The red colour will go away once the user picks another, legal spot.