fix: DNRY
This is a minor edit. I made my code more compact. Rather than using an if /else statement for determining if the winner is X or O, I made an if/else statement to determine the value of a variable. This variable, whatever its value, can be used to check for a winner. This prevents the code from having to repeat itself.