fix: all wins recognised
This code recognises a win by using prime factorisation: each state of the board is given a unique numerical identifier. This is fun, but comes with some down sides. Most notably, earlier drafts of this code did not know how to handle redundancy. So, a player could win but, if s/he was unfortunate enough to make a redundant move, then the win would not be recognised.

This commit fixes this issue. It does this through the use of counters: if a player makes 4 moves then, if they win, s/he will have one redundant step; if the player makes 5 moves then if s/he wins s/he will have made two redundant moves (a player cannot make more than 5 moves, since there are  only 9 spots on the board). The code then deals with these potential redundancies accordingly, by looking within them for possible sub-groups that, when multiplied together, equal the numerical identifier of one of the non-redundant winning states of the board.