fix: if win, end game
I have made it so that if a player wins, the game is clearly ended. Users are no longer able to select new squares to occupy them with an X or O. Now, if a board is selected after the game has been won, it will just shake (which is meant to show the user(s) that there has been an error).