feat/fix: draw animation
If all the squares on the tic-tac-toe board are used up, and we have no winner, then the game ends in a draw. In this commit, the game now recognises a draw through an animation of the game board shaking. This should signal to the user(s) that something has changed.