feat: add GUI win animation
I added a winning animation for the graphic user interface. Now, when the game is over, animated text will appear on the screen that informs users which player won (if there is a player that won) or that it is a draw (if the game ends in a draw).

These text animations co-occur with the other end-game animations added in previous commits.