minor: test
I am having issues with storing variables between pages. According to Stack Overflow this issue could be caused by my local-server vim plugin, so I am pushing my repo to origin to see if the issue persists with GitHub pages. If it doesn't, then I know the issue is with my local server; if it does, then I know the issue is with my code.