feat: input
I have made it so that when names are typed into input elements, there is a js event listener that updates the name of the corresponding player for each new input.

If the name is valid, a green checkmark will be shown next to the input element. If the name is invalid (i.e. it is identical to the other player's name) then a red X will be visible next to the input element. Both the checkmark and the X shake as a result of a CSS animation.