feat: chose if player is ai or not
Now the user can determine if the player is an AI or not. The actual AI itself still does not work -- that will come with later commits -- but at least now the Player objects have a way of responding to radio buttons and updating according to the user selection (i.e. it updates according to whether the user wants the player to be an AI).