fix: all players can win
As described in earlier commits, I ran across a bug that made it so that
only one player could win the game (for the other player, wins were not
recognized). By going back and looking at previous code, I have managed
to locate the source of this error: a minor typo in js/intro.js, which
assigned the wrong values to the wrong players. These values may have
been used to ID players later on in the code, so since they were
incorrectly assigned this lead to false negatives. But regardless of the
specific mechanisms at play, the problem is now fixed.