fix: win without restart
In my last commit, I purported to have fixed the bug that inhibits one of two players from winning the game. But I missed a crucial detail: the fix only worked after a game has been played once through and then refreshed through the "restart game" button.

The reason for this seems to be that the button set all the relevant variables/properties back to values different from the default values of the game. I managed to pinpoint the specific ways in which these two states of the game diverged, and correct the default state of the game so that it is the same as the values set by the 'restart' button. This appears to have fixed the issue, since both players are capable of winning a game now, even on their first go.