fix: change X/O, end message
I have changed the GUI so that X starts first (as is the convention). This has the added benefit of making it so that the X and Os displayed on the browser window are consistent with the placement of X and Os within the game.board array (they weren't before, which was a problem, because it risks making the game logic come apart from its graphics).

I have also changed the styling of the win messages (before they would occasionally just outside the browser window, but now they always appear just below the tic-tac-toe board).