feat: find win recursively
I am still in the process of making the AI. My approach so far is to do something similar to what is done to find a redundant win. I am looking for the available squares, and seeing all the combos that lead to a win. From there, I see which particular square appears the most within these possibilities, and then selecting that square. But there is still much debugging to do.